Restructure code.
