+++ lib/screens/settings/document_presets.dart
@@ -23,9 +23,9 @@ import 'package:nahpu/services/providers/specimens.dart';
 import 'package:nahpu/services/providers/sites.dart';
 import 'package:nahpu/services/providers/collevents.dart';
 import 'package:nahpu/services/providers/narrative.dart';
-import 'package:nahpu/services/template_service.dart';
+import 'package:nahpu/services/templates/template_service.dart';
 import 'package:nahpu/services/types/export.dart';
-import 'package:nahpu/services/template_settings_services.dart';
+import 'package:nahpu/services/templates/template_settings_services.dart';
 import 'package:nahpu/services/print_specimen_table_columns.dart';
 import 'package:nahpu/services/platform_services.dart';
 import 'package:nahpu/services/export/export_document.dart';
