+++ lib/screens/templates/template_editor_screen.dart
@@ -11,9 +11,9 @@ import 'package:nahpu/screens/templates/template_editor_math.dart';
 import 'package:nahpu/screens/templates/template_fonts.dart';
 import 'package:nahpu/screens/templates/template_model.dart';
 import 'package:nahpu/services/types/export.dart';
-import 'package:nahpu/services/template_settings_services.dart';
-import 'package:nahpu/services/template_service.dart';
-import 'package:nahpu/services/template_editor_service.dart';
+import 'package:nahpu/services/templates/template_settings_services.dart';
+import 'package:nahpu/services/templates/template_service.dart';
+import 'package:nahpu/services/templates/template_editor_service.dart';
 import 'package:nahpu/services/export/document_writer.dart';
 import 'package:nahpu/screens/templates/components/layout/template_editor_scaffold.dart';
 import 'package:nahpu/screens/templates/components/properties/text_element_editor.dart';
