+++ lib/services/export/document_writer.dart
@@ -8,8 +8,8 @@ import 'package:flutter/services.dart' show rootBundle, AssetManifest;
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:nahpu/screens/templates/template_model.dart';
 import 'package:nahpu/services/database/database.dart';
-import 'package:nahpu/services/template_service.dart';
-import 'package:nahpu/services/template_settings_services.dart';
+import 'package:nahpu/services/templates/template_service.dart';
+import 'package:nahpu/services/templates/template_settings_services.dart';
 import 'package:nahpu/services/providers/database.dart';
 import 'package:nahpu/services/export/dynamic_record_exporter.dart';
 import 'package:nahpu/services/collevent_services.dart';
