+++ lib/services/templates/template_image_service.dart
